add toggle for listsort

+++ response/list.js
@@ -3,6 +3,8 @@ module.exports = {
 	description: '',
 	execute(message) {
 		if (message.author.id == '765287593762881616') return;
+		const srvconfig = client.settings.get(message.guild.id);
+		if (srvconfig.listsort == 'false') return;
 		message.delete();
 		const count = message.content.split(/\n+/)[0];
 		const players = message.content.replace(`${count}\n\`\`\`\n`, '').replace('\n```', '').split(/, /).sort();
